refactor: 在保单状态变换报文中将日期替换为时刻
